cache intercepter definition to renderer class
